update: main.cpp linked.cpp linked.hpp - add description - update s.t. service time will be generated in the range of 0.1-10 - add struct statistics served as return struct from scheduling algorithms for result display

+++ project-2/linked.cpp
@@ -12,7 +12,7 @@ Job::Job()
 
 //Makes the Job Class with parameters
 //Param: a->arrival, b->service, nam-> name of service
-Job::Job(int a, int b, int s, std::string nam = "None")
+Job::Job(int a, double b, int s, std::string nam = "None")
 {
 	arrivalTime = a;
 	serviceTime = b;
